This text editor print the contents of a file to the screen, page by page. Each page consist of 10 lines.
My work in this project is make a command-based text editor. This text editor print the contents of a file to the screen, page by page. Each page consist of 10 lines.

My program can keep the contents of the file in certain data structures, and can support certain commands.

In order to keep a single text line, I used the C++ string class in order to simplify string operations. However, the lines kept in a linked list structure. Each line stored in a separate node.

Also, I utilized a stack in order to keep track of all actions carried out, so I can undo actions as the user tells the program to. I also defined a structure of my own in order to store relevant data on previous commands.

Commands:
While my program is running, it repeatedly prints the contents of the current file and ask the user what user would like do. The following are the valid commands the user can enter:
-open filename
-save filename
-insert n text
-delete n
-move n m
-replace n text
-next
-prev
-undo